services/metrics-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0

# Redis
redis==5.0.1

# gRPC
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
